# application/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "A comprehensive API for managing a mechanic shop including customers, mechanics, service tickets, and inventory"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "JWT Authorization header using the Bearer scheme. Example: 'Bearer {token}'"
paths:
...
  /inventory:
    post:
      tags:
        - Inventory
      summary: "Create a new inventory part"
      description: "Add a new part to the inventory system"
      parameters:
        - in: "body"
          name: "body"
          description: "Inventory part information"
          required: true
          schema:
            $ref: "#/definitions/CreateInventoryPayload"
      responses:
        201:
          description: "Inventory part created successfully"
          schema:
            $ref: "#/definitions/InventoryResponse"
          examples:
            application/json:
              id: 1
              name: "Oil Filter"
              price: 12.99
        400:
          description: "Invalid input"
          schema:
            $ref: "#/definitions/ErrorResponse"

    get:
      tags:
        - Inventory
      summary: "Returns all Inventory Parts"
      description: "Endpoint to retrieve a list of all inventory parts."
      responses:
        200:
          description: "Retrieved Inventory Parts Successfully"
          schema:
            $ref: "#/definitions/AllInventory"
  /inventory/{inventory_id}:
    get:
      tags:
        - Inventory
      summary: "Get a specific inventory part by ID"
      description: "Retrieve detailed information about a specific inventory part"
      parameters:
        - in: "path"
          name: "inventory_id"
          type: "integer"
          required: true
          description: "ID of the inventory part to retrieve"
      responses:
        200:
          description: "Inventory part retrieved successfully"
          schema:
            $ref: "#/definitions/InventoryResponse"
        404:
          description: "Inventory part not found"
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      tags:
        - Inventory
      summary: "Update an inventory part"
      description: "Update an existing inventory part's information"
      parameters:
        - in: "path"
          name: "inventory_id"
          type: "integer"
          required: true
          description: "ID of the inventory part to update"
        - in: "body"
          name: "body"
          description: "Updated inventory part information"
          required: true
          schema:
            $ref: "#/definitions/UpdateInventoryPayload"
      responses:
        200:
          description: "Inventory part updated successfully"
          schema:
            $ref: "#/definitions/InventoryResponse"
        400:
          description: "Invalid input"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Inventory part not found"
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      tags:
        - Inventory
      summary: "Delete an inventory part"
      description: "Remove an inventory part from the system"
      parameters:
        - in: "path"
          name: "inventory_id"
          type: "integer"
          required: true
          description: "ID of the inventory part to delete"
      responses:
        200:
          description: "Inventory part deleted successfully"
          schema:
            $ref: "#/definitions/DeleteResponse"
        404:
          description: "Inventory part not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
definitions:
...
  CreateInventoryPayload:
    type: "object"
    properties:
      name:
        type: "string"
        example: "Oil Filter"
      price:
        type: "number"
        format: "float"
        example: 12.99
    required:
      - name
      - price
  UpdateInventoryPayload:
    type: "object"
    properties:
      name:
        type: "string"
        example: "Premium Oil Filter"
      price:
        type: "number"
        format: "float"
        example: 15.99
    required:
      - name
      - price
  InventoryResponse:
    type: "object"
    properties:
      id:
        type: "integer"
        example: 1
      name:
        type: "string"
        example: "Oil Filter"
      price:
        type: "number"
        example: 12.99
  AllInventory:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
          example: 1
        name:
          type: "string"
          example: "Oil Filter"
        price:
          type: "number"
          example: 12.99

  # General Definitions
  DeleteResponse:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Successfully deleted"
  ErrorResponse:
    type: "object"
    properties:
      error:
        type: "string"
        example: "An error occurred"
      errors:
        type: "object"